ajout du user_id

+++ src/TaskGateway.php
@@ -29,13 +29,15 @@ class TaskGateway
         return $data;
     }
 
-    public function get(string $id)
+    public function getForUser(int $user_id, string $id): array |false
     {
-        $sql = "SELECT * FROM task WHERE id = :id";
+        $sql = "SELECT * FROM task WHERE id = :id AND user_id = :user_id";
+
 
         $stmt = $this->conn->prepare($sql);
 
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
 
         $stmt->execute();
 
@@ -49,9 +51,9 @@ class TaskGateway
         return $data;
     }
 
-    public function create(array $data): string
+    public function createForUser(int $user_id, array $data): string
     {
-        $sql = "INSERT INTO task(name,priority,is_completed) VALUES (:name, :priority, :is_completed)";
+        $sql = "INSERT INTO task(name,priority,is_completed, user_id) VALUES (:name, :priority, :is_completed,:user_id)";
 
         $stmt = $this->conn->prepare($sql);
 
@@ -65,13 +67,15 @@ class TaskGateway
 
         $stmt->bindValue(":is_completed", $data["is_completed"] ?? false, PDO::PARAM_BOOL);
 
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
+
         $stmt->execute();
 
         return $this->conn->lastInsertId();
 
     }
 
-    public function update(string $id, array $data): int
+    public function updateForUser(int $user_id, string $id, array $data): int
     {
         $fields = [];
 
@@ -105,11 +109,12 @@ class TaskGateway
                 return "$value = :$value";
             }, array_keys($fields));
 
-            $sql = "UPDATE task" . " SET " . implode(",", $sets) . " WHERE id = :id";
+            $sql = "UPDATE task" . " SET " . implode(",", $sets) . " WHERE id = :id" . "AND user_id = :user_id";
 
             $stmt = $this->conn->prepare($sql);
 
             $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+            $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
 
             foreach ($fields as $name => $values) {
                 $stmt->bindValue(":$name", $values[0], $values[1]);
@@ -122,11 +127,12 @@ class TaskGateway
 
     }
 
-    public function delete(string $id): int
+    public function deleteForUser(int $user_id, string $id): int
     {
-        $sql = "DELETE FROM task WHERE id= :id";
+        $sql = "DELETE FROM task WHERE id= :id AND user_id = :user_id";
         $stmt = $this->conn->prepare($sql);
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
         $stmt->execute();
         return $stmt->rowCount();
     }
